attempt at loading cols

+++ Pumpkin Knight/Motor2D/j1Player.h
@@ -20,11 +20,11 @@ public:
 	bool Start();
 	bool Update(float dt);
 	bool CleanUp();
-	int getX()
+	float getX()
 	{
 		return pos.x;
 	}
-	int getY()
+	float getY()
 	{
 		return pos.y;
 	}
